projects/smart-claims-processor/requirements.txt
# Core LangGraph + LangChain
langgraph>=0.2.50
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# CrewAI (for fraud detection crew)
crewai>=0.80.0
crewai-tools>=0.15.0

# LLM Provider
google-generativeai>=0.8.0

# Data validation
pydantic>=2.6.0

# Web API (HITL review server)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Dashboard
streamlit>=1.40.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
